removed all input related functions from cli_controller and added them to input_handler
