+++ module/cli_controller.py
@@ -1,11 +1,9 @@
-import threading
-import queue
-
 import common.frames
 from client.comm import BaseComm
 
 from module.command_node import CommandNode
 from module.command_tree_generator import load_commands
+from module.input_handler import input_handler
 
 
 class CLIController:
@@ -16,9 +14,6 @@ class CLIController:
         """
         self.comm = comm
         self.root_node = CommandNode("ROOT")
-
-        self.input_queue = queue.Queue()
-        self.input_thread = threading.Thread()
 
         # These function as preserved keywords, do not use these names in commands
 
@@ -35,6 +30,8 @@ class CLIController:
         self.command_file_list = command_file_list
         self.load_tree()
         self.current_node = self.root_node
+
+        self.input_handler = input_handler(self)
 
         self.stopped = False
 
@@ -68,86 +65,20 @@ class CLIController:
         else:
             print("\tThis function requires no parameters and has no children")
 
-    @staticmethod
-    def ask_input(input_queue: queue.Queue, string=""):
-        """
-        Starts a new thread asking the user for input and writes this input to the given input_queue
-        Optional string for input
-        """
-        i = input(string)
-        input_queue.put(i)
-
-    """
-    Starts a new thread to make nonblocking input possible. And get the current location, after restart this is always just root
-    """
-    def start_thread(self):
-        s = self.make_path_string(self.current_node.get_branch_names()) + " "
-        self.input_thread = threading.Thread(target=self.ask_input, args=(self.input_queue, s))
-        self.input_thread.daemon = True
-        self.input_thread.start()
-
-    """
-    Go back one node in the tree structure. You cant go back when in root
-    """
     def go_back_in_tree(self):
+        """
+        Go back one node in the tree structure. You cant go back when in root
+        """
         if self.current_node.parent:
             self.current_node = self.current_node.parent
 
-    """
-    Go to root in tree structure
-    """
+
     def go_to_root(self):
+        """
+        Go to root in tree structure
+        """
         self.current_node = self.root_node
-
-    """
-    Handles all non-global commands. Returns false if failed or if a function has been executed(in this case no other commands can be executed after).
-    Returns true if another command can be executed after this one
-    """
-    def handle_nonglobal_commands(self, user_word, user_command_list):
-        if user_word.upper() in self.current_node.keys():
-            self.current_node = self.current_node[user_word.upper()]
-            return True
-
-        elif len(self.current_node.parameter_list) > 0:
-            print("\tCommand called with {} parameters: {}".format(
-                len(user_command_list[user_command_list.index(user_word):]),
-                "(" + ",".join(user_command_list[user_command_list.index(user_word):]) + ")"
-            ))
-
-        elif user_word:
-            print("\tCommand {} not found, type \"help\" for possible commands.".format(user_word))
-        return False
-
-    """
-    Execute a command depending on text entered
-    """
-    def handle_new_input(self):
-        user_command_list = self.input_queue.get().split(" ")
-        for user_word in user_command_list:
-
-            # Step 1: Check for global commands
-            if user_word.upper() in self.global_commands.keys():
-                self.global_commands[user_word.upper()]()
-                    
-            # Step 2: Check for location(in tree structure) specific commands
-            else:
-                if not self.handle_nonglobal_commands(user_word, user_command_list):
-                    break
                 
-    """
-    Starts an infinite loop (until exit command is called) which polls for input
-    """
-    def run_cli(self):
-        """
-        Starts thread asking for input if it is currently not and the input_queue is not filled.
-        Otherwise processes items in the input_queue.
-        """
-        if not self.input_thread.isAlive() and self.input_queue.empty():
-            self.start_thread()
-        elif not self.input_queue.empty():
-            self.handle_new_input()
-
-
     def process(self):
         """
         Main loop of the module
@@ -155,15 +86,13 @@ class CLIController:
         while self.comm.has_data():
             frame = self.comm.get_data()
 
-        self.run_cli()
-
-
+        self.input_handler.check_input()
 
     def stop(self):
         """
         Stops the CLIController
         """
-        if self.input_thread:
-            self.input_thread.join()
+        if self.input_handler.input_thread:
+            self.input_handler.input_thread.join()
         self.comm.stop()
         self.stopped = True
